+++ features/products/actions/index.ts
@@ -1,0 +1,3 @@
+export * from "./create-product.action";
+export * from "./update-product.action";
+export * from "./delete-product.action";