+++ features/products/index.ts
@@ -1,0 +1,7 @@
+export * from "./types/product.types";
+export * from "./validations/product.schema";
+export * from "./services/product.service";
+export * from "./queries";
+export * from "./mutations";
+export * from "./actions";
+export * from "./components";