+++ features/products/queries/index.ts
@@ -1,0 +1,2 @@
+export * from "./list-products";
+export * from "./get-product";